+++ include/poica/interface.h
@@ -30,54 +30,37 @@
 #include <poica/private/prefix.h>
 
 #include <poica/interface/gen/i_obj.h>
-#include <poica/interface/gen/methods_ptrs.h>
+#include <poica/interface/gen/method_tables.h>
+#include <poica/interface/i_obj_utils.h>
 #include <poica/interface/method.h>
+
+#include <poica/interface/checks/is_no_mut_self_methods.h>
+#include <poica/interface/checks/is_no_self_methods.h>
+#include <poica/interface/checks/is_no_static_methods.h>
 
 #include <boost/preprocessor.hpp>
 
 #ifdef POICA_USE_PREFIX
-
-#define poicaInterface        POICA_P_INTERFACE
-#define poicaVTable           POICA_P_VTABLE
-#define poicaVTableImpl       POICA_P_VTABLE_IMPL
-#define poicaIObjCall         POICA_P_I_OBJ_CALL
-#define poicaIObjCallWithArgs POICA_P_I_OBJ_CALL_WITH_ARGS
-#define poicaInitIObj         POICA_P_INIT_I_OBJ
-
+#define poicaInterface POICA_P_INTERFACE
 #else
-
-#define interface        POICA_P_INTERFACE
-#define VTable           POICA_P_VTABLE
-#define VTableImpl       POICA_P_VTABLE_IMPL
-#define iObjCall         POICA_P_I_OBJ_CALL
-#define iObjCallWithArgs POICA_P_I_OBJ_CALL_WITH_ARGS
-#define initIObj         POICA_P_INIT_I_OBJ
-
+#define interface POICA_P_INTERFACE
 #endif
 
-#define POICA_P_INTERFACE(name, functions)                                     \
-    typedef struct POICA_P_VTABLE(name) {                                      \
-        POICA_P_INTERFACE_GEN_METHODS_PTRS(functions)                          \
-    } POICA_P_VTABLE(name);                                                    \
+#define POICA_P_INTERFACE(name, methods)                                       \
+    BOOST_PP_IF(POICA_P_INTERFACE_IS_NO_STATIC_METHODS(methods),               \
+                BOOST_PP_EMPTY(),                                              \
+                POICA_P_INTERFACE_GEN_METHOD_TABLES_STATIC(name, methods);)    \
                                                                                \
-    POICA_P_GEN_I_OBJ(name);                                                   \
-    POICA_P_GEN_MUT_I_OBJ(name);                                               \
+    BOOST_PP_IF(POICA_P_INTERFACE_IS_NO_SELF_METHODS(methods),                 \
+                BOOST_PP_EMPTY(),                                              \
+                POICA_P_INTERFACE_GEN_METHOD_TABLES_SELF(name, methods);       \
+                POICA_P_GEN_I_OBJ(name);)                                      \
+                                                                               \
+    BOOST_PP_IF(POICA_P_INTERFACE_IS_NO_MUT_SELF_METHODS(methods),             \
+                BOOST_PP_EMPTY(),                                              \
+                POICA_P_INTERFACE_GEN_METHOD_TABLES_MUT_SELF(name, methods);   \
+                POICA_P_GEN_MUT_I_OBJ(name);)                                  \
                                                                                \
     POICA_FORCE_SEMICOLON
-
-#define POICA_P_VTABLE(interface_name)                                         \
-    POICA_P_PREFIX(BOOST_PP_CAT(interface_name, _VTable))
-
-#define POICA_P_I_OBJ_CALL(i_obj, method) (i_obj).vtable->method(&(i_obj))
-
-#define POICA_P_I_OBJ_CALL_WITH_ARGS(i_obj, method, ...)                       \
-    (i_obj).vtable.method(&(i_obj), __VA_ARGS__)
-
-#define POICA_P_VTABLE_IMPL(interface_name, implementer_name)                  \
-    BOOST_PP_CAT(POICA_P_VTABLE(interface_name), implementer_name)
-
-#define POICA_P_INIT_I_OBJ(i_obj, new_self, new_vtable)                        \
-    (i_obj).self = (new_self);                                                 \
-    (i_obj).vtable = (new_vtable);
 
 #endif // POICA_INTERFACE_H